Fix linux deploys failing on desktop file validation
The `SingleMainWindow` key added to `osu!.desktop` in #145 has been
added to the Desktop Entry Specification in version 1.5 [1].
`appimagetool` was using an older version of `desktop-file-validate`
that only understood version 1.0 [2], thus leading to deployment
failure.

To fix, update `appimagetool` from release 13 to the continuous build to
receive the upstream fixes from [2], as well as explicitly specify the
Desktop Entry Specification version in the `osu!.desktop` file template.

[1]: https://specifications.freedesktop.org/desktop-entry-spec/desktop-entry-spec-latest.html#idm45828204418512
[2]: https://github.com/AppImage/AppImageKit/issues/1171