Fix a silly path issue
So when I converted this to use local tools instead of global tools, it
turns out that I forgot to account for the *working directory* of the
commands. It was being set to the osu!.Desktop solution folder, rather
than osu-deploy, resulting in an error that the vpk tool couldnt be
found.

This fixes that, and since the commands process absolute paths anyways,
it doesnt really change anything.